requirements.txt
pytest==3.9.1
flake8==3.5.0
numpy==1.15.2
tensorflow==1.11.0
scipy==1.1.0
Pillow==5.3.0
requests==2.20.0
googledrivedownloader==0.3
tqdm==4.28.0
packaging==21.3
psycopg2==2.9.3
